LRSD-10482 Added TeamMembers.ts service file to place addContactRoleNameByEmailByProject and deleteContactRoleNameByEmailByProject functions that were moved from a less relevant file.
